이미지 사이즈 변경

+++ src/components/ImageUpload.js
@@ -11,11 +11,29 @@ const ImageUpload = ({ ondownloadURL, name }) => {
   const handleImageChange = async (e) => {
     const file = e.target.files[0];
     if (file && file.type.startsWith('image/')) {
-      setImage(file);
-      if (url) {
-        await deleteExistingImage(url);
-      }
-      await handleSubmit(file);
+      const img = new Image();
+      img.src = URL.createObjectURL(file);
+      img.onload = async () => {
+        if (img.width !== 1024 || img.height !== 576) {
+          if (window.confirm('이미지 크기가 1024*576이 아닙니다. 수정해드릴까요?')) {
+            const resizedImage = await resizeImage(img, 1024, 576, file.name);
+            setImage(resizedImage);
+            if (url) {
+              await deleteExistingImage(url);
+            }
+            await handleSubmit(resizedImage);
+          } else {
+            alert('이미지를 업로드할 수 없습니다.');
+            setImage(null);
+          }
+        } else {
+          setImage(file);
+          if (url) {
+            await deleteExistingImage(url);
+          }
+          await handleSubmit(file);
+        }
+      };
     } else {
       alert('Please select a valid image file');
       setImage(null);
@@ -52,6 +70,32 @@ const ImageUpload = ({ ondownloadURL, name }) => {
       console.error('Error uploading image: ', error);
       alert('Error uploading image');
     }
+  };
+
+  const resizeImage = (img, width, height, originalFileName) => {
+    return new Promise((resolve) => {
+      const canvas = document.createElement('canvas');
+      const ctx = canvas.getContext('2d');
+      canvas.width = width;
+      canvas.height = height;
+      ctx.fillStyle = 'rgba(0, 0, 0, 0)'; // 투명색으로 채우기
+      ctx.fillRect(0, 0, width, height);
+
+      const aspect = img.width / img.height;
+      let newWidth, newHeight;
+      if (aspect > width / height) {
+        newWidth = width;
+        newHeight = width / aspect;
+      } else {
+        newWidth = height * aspect;
+        newHeight = height;
+      }
+      ctx.drawImage(img, (width - newWidth) / 2, (height - newHeight) / 2, newWidth, newHeight);
+
+      canvas.toBlob((blob) => {
+        resolve(new File([blob], originalFileName, { type: 'image/png' }));
+      }, 'image/png');
+    });
   };
 
   return (
